Used passport-jwt to delete user posts

+++ controllers/api/v1/posts_api.js
@@ -1,5 +1,6 @@
 const Posts = require('../../../models/posts'); //Importing Post Schema
 const Comment = require('../../../models/comment');   //Importing the Comment Schema
+const jwt = require('jsonwebtoken');            //JWT Library
 
 module.exports.index = async (req, res) => {    //.index for listing the data
 
@@ -28,33 +29,23 @@ module.exports.destroy = async function(req, res){
         //post.user is the user from posts schema which returns the user id.
         /*Comparing if the currently authorized user id is same as the post.user 
         So that, deleting permission for other users are restricted*/  
-        // if(post.user == req.user.id){
+        if(post.user == req.user.id){
             post.remove();          //Deleting the post
 
             //Deleting all/ the Comments of the post
             await Comment.deleteMany({post: req.params.id});
 
-            // if (req.xhr){
-            //     return res.status(200).json({
-            //         data: {
-            //             post_id: req.params.id
-            //         },
-            //         message: "Post deleted"
-            //     });
-            // }
-            //req.flash('success', 'Post Deleted');
             return res.status(200).json({
                 message: "Post and associated comments Deleted"
             });
-        // }
-        // else{
-        //     req.flash('error', 'You cannot delete this post!');
-        //     return res.redirect('back');
-        // }
+        }
+        else{
+            return res.status(401).json({
+                message: "You're Unauthorised to delete this posts"
+            });
+        }
         
     } catch(err){
-        console.log(err);
-        //req.flash('error', err);
         return res.status(500).json({
             message: "Internal Server Error !!"
         });
